Crate solana_net_utils

The net_utils module assists with networking

Modules

multihomed_sockets

sockets

token_bucket

This module contains TokenBucket, which provides ability to limit rate of certain events, while allowing bursts through. KeyedRateLimiter allows to rate-limit multiple keyed items, such as connections.

Structs

SocketConfigDeprecated

UdpSocketPair

A data type representing a public Udp socket

Constants

DEFAULT_IP_ECHO_SERVER_THREADS

MAX_PORT_COUNT_PER_MESSAGE

MINIMUM_IP_ECHO_SERVER_THREADS

MINIMUM_VALIDATOR_PORT_RANGE_WIDTH

VALIDATOR_PORT_RANGE

Functions

bind_commonDeprecated

binds both a UdpSocket and a TcpListener

bind_common_in_range_with_configDeprecated

Find a port in the given range with a socket config that is available for both TCP and UDP

bind_common_with_configDeprecated

binds both a UdpSocket and a TcpListener on the same port

bind_in_range

bind_in_range_with_configDeprecated

bind_more_with_configDeprecated

bind_toDeprecated

bind_to_localhost

bind_to_unspecified

bind_to_with_configDeprecated

bind_to_with_config_non_blockingDeprecated

bind_two_in_range_with_offsetDeprecated

bind_two_in_range_with_offset_and_configDeprecated

bind_with_any_port_with_configDeprecated

find_available_port_in_range

Searches for an open port on a given binding ip_addr in the provided range.

find_available_ports_in_range

Searches for several ports on a given binding ip_addr in the provided range.

get_cluster_shred_version

Retrieves cluster shred version from Entrypoint address provided.

get_cluster_shred_version_with_binding

Retrieves cluster shred version from Entrypoint address provided, binds client-side socket to the IP provided.

get_public_ip_addrDeprecated

Determine the public IP address of this machine by asking an ip_echo_server at the given address.

get_public_ip_addr_with_binding

Determine the public IP address of this machine by asking an ip_echo_server at the given address. This function will bind to the provided bind_addreess.

ip_echo_server

Starts a simple TCP server that echos the IP address of any peer that connects Used by functions like |get_public_ip_addr| and |get_cluster_shred_version|

is_host

is_host_port

multi_bind_in_range_with_configDeprecated

binds num sockets to the same port in a range with config

parse_host

parse_host_port

parse_port_or_addr

parse_port_range

verify_all_reachable_tcp

Checks if all of the provided TCP ports are reachable by the machine at ip_echo_server_addr. Tests must complete within timeout provided. Tests will run concurrently when possible, using up to 64 threads for IO. This function assumes that all sockets are bound to the same IP, and will panic otherwise.

verify_all_reachable_udp

Checks if all of the provided UDP ports are reachable by the machine at ip_echo_server_addr. Tests must complete within timeout provided. Tests will run concurrently when possible, using up to 64 threads for IO. This function assumes that all sockets are bound to the same IP, and will panic otherwise

Type Aliases

IpEchoServer

PortRange