solana_net_utils

Function find_available_ports_in_range

pub fn find_available_ports_in_range<const N: usize>(
    ip_addr: IpAddr,
    range: PortRange,
) -> Result<[u16; N]>

Searches for several ports on a given binding ip_addr in the provided range.

This will start at a random point in the range provided, and search sequentially. If it can not find anything, an Error is returned.