solana_net_utils

Module sockets

Structs

SocketConfiguration

Functions

bind_common_in_range_with_config

Find a port in the given range with a socket config that is available for both TCP and UDP

bind_common_with_config

binds both a UdpSocket and a TcpListener on the same port

bind_gossip_port_in_range

bind_in_range_with_config

bind_more_with_config

bind_to

bind_to_localhost_unique

Bind a UdpSocket to a unique port.

bind_to_with_config

bind_two_in_range_with_offset_and_config

bind_with_any_port_with_configDeprecated

localhost_port_range_for_tests

Retrieve a free 25-port slice for unit tests

multi_bind_in_range_with_config

binds num sockets to the same port in a range with config

unique_port_range_for_tests

When running under nextest, this will try to provide a unique slice of port numbers (assuming no other nextest processes are running on the same host) based on NEXTEST_TEST_GLOBAL_SLOT variable The port ranges will be reused following nextest logic.