Crate solana_net_utils

The net_utils module assists with networking

Structs

UdpSocketPair

A data type representing a public Udp socket

Constants

MAX_PORT_COUNT_PER_MESSAGE

MINIMUM_VALIDATOR_PORT_RANGE_WIDTH

VALIDATOR_PORT_RANGE

Functions

bind_common

bind_common_in_range

bind_in_range

bind_to

bind_two_in_range_with_offset

bind_with_any_port

find_available_port_in_range

get_cluster_shred_version

get_public_ip_addr

Determine the public IP address of this machine by asking an ip_echo_server at the given address

ip_echo_server

Starts a simple TCP server on the given port that echos the IP address of any peer that connects. Used by |get_public_ip_addr|

is_host

is_host_port

multi_bind_in_range

parse_host

parse_host_port

parse_port_or_addr

parse_port_range

verify_reachable_ports

Type Aliases

IpEchoServer

PortRange